requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-crispy-forms==1.14.0
django-fontawesome-5==1.0.18
django-froala-editor==4.0.16
django-js-asset==2.1.0
django-recaptcha
django-honeypot
django-simple-sso==1.2.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.8.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
yarl==1.9.2
